@@ -94,15 +94,13 @@
     "@angular/compiler-cli": "^20.3.2",
     "@angular/language-service": "^20.3.2",
     "@eslint/eslintrc": "^3.1.0",
-    "@eslint/js": "^9.8.0",
+    "@eslint/js": "^9.36.0",
     "@schematics/angular": "^20.3.3",
     "@types/gulp": "4.0.6",
     "@types/highlight.js": "^9.12.4",
     "@types/jasmine": "~3.10.0",
     "@types/jasminewd2": "2.0.3",
     "@types/node": "^20.19.17",
-    "@typescript-eslint/eslint-plugin": "^8.0.0",
-    "@typescript-eslint/parser": "^8.0.0",
     "angular-cli-ghpages": "0.5.0",
     "angular-eslint": "~20.3.0",
     "browserstack-local": "^1.3.7",
@@ -110,8 +108,8 @@
     "conventional-changelog-cli": "1.3.4",
     "del": "3.0.0",
     "doc-prsr": "2.2.0",
-    "eslint": "^9.8.0",
-    "eslint-config-prettier": "^9.1.0",
+    "eslint": "^9.36.0",
+    "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-rxjs-x": "^0.7.7",
     "fs-extra": "^8.1.0",
     "globals": "^15.9.0",
@@ -123,7 +121,7 @@
     "jasmine-core": "~3.10.0",
     "jasmine-fail-fast": "2.0.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.0",
+    "karma": "~6.4.4",
     "karma-browserstack-launcher": "^1.6.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
@@ -139,7 +137,8 @@
     "stylelint-config-prettier": "^9.0.3",
     "ts-node": "^10.4.0",
     "typedoc": "^0.12.0",
-    "typescript": "~5.9.2"
+    "typescript": "~5.9.2",
+    "typescript-eslint": "^8.45.0"
   },
   "schematics": "./tools/dev-schematics/dist/collection.json"
 }
