@@ -1,6 +1,5 @@
 import { OnInit, ChangeDetectorRef, Directive, ElementRef, HostBinding, Input, OnDestroy } from '@angular/core';
-import { Subject } from 'rxjs';
-import { distinctUntilChanged, map, takeUntil } from 'rxjs/operators';
+import { distinctUntilChanged, map, Subject, takeUntil } from 'rxjs';
 import { ComponentsListService } from './components-list.service';
 
 @Directive({
