@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { map, Observable } from 'rxjs';
 
 import { NbAuthService, NbAuthJWTToken, NbAuthToken } from '@nebular/auth';
 import { NbRoleProvider } from '@nebular/security';
